pyproject.toml
[project]
name = "django-2fa"
version = {from = "django_2fa/__init__.py"}
description = "2 Factor authenication application for Django"
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "django-fernet-encrypted-fields~=0.0",
    "pyotp~=2.6",
]
requires-python = ">=3.7"
dynamic = ["classifiers", "version"]
license = {text = "MIT"}
readme = "README.md"

[project.urls]
homepage = "https://github.com/neutron-sync/django-2fa"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.scripts]
publish = "pdm publish --username pizzapanther"
